+++ plugins/destination/postgresql/client/types.go
@@ -9,6 +9,8 @@ func (c *Client) SchemaTypeToPg(t arrow.DataType) string {
 	switch c.pgType {
 	case pgTypeCockroachDB:
 		return pgarrow.ArrowToCockroach(t)
+	case pgTypeCrateDB:
+		return pgarrow.ArrowToCrateDB(t)
 	default:
 		return pgarrow.ArrowToPg10(t)
 	}
@@ -18,6 +20,8 @@ func (c *Client) PgToSchemaType(t string) arrow.DataType {
 	switch c.pgType {
 	case pgTypeCockroachDB:
 		return pgarrow.CockroachToArrow(t)
+	case pgTypeCrateDB:
+		return pgarrow.CrateDBToArrow(t)
 	default:
 		return pgarrow.Pg10ToArrow(t)
 	}